+++ imcrvtip/CandidateWindow.h
@@ -51,7 +51,7 @@ public:
 	void _Redraw();
 	void _SetText(const std::wstring &text, BOOL fixed, BOOL showcandlist, BOOL showreg);
 	void _GetPrecedingText(std::wstring *text);
-	void _DeletePrecedingText(int delete_count);
+	void _DeletePrecedingText(size_t delete_count);
 	void _PreEnd();
 	void _End();
 	void _UpdateComp();
